Add CRUD for user assets & transactions with tests
Implemented full CRUD operations for user assets and user asset transactions, including API endpoints, request/response models, handlers, validators, mapping profiles, and service abstractions. Updated repository interfaces and DI setup. Added comprehensive unit tests for all new endpoints and removed placeholder tests. Test project now references the API project.

+++ backend/CryptoPortfolio/Domain/Repositories/IUserAssetTransactionRepository.cs
@@ -6,7 +6,7 @@ namespace CryptoPorfolio.Domain.Repositories
     public interface IUserAssetTransactionRepository : IDomainRepository
     {
         Task<IEnumerable<UserAssetTransaction>> GetUserTransactionsAsync(
-            int userAssetId,
+            long userAssetId,
             CancellationToken cancellationToken = default);
 
         Task<UserAssetTransaction?> GetByIdAsync(long id, CancellationToken cancellationToken = default);
